Enum foundry_evm_core::abi::HardhatConsole::HardhatConsoleCalls

pub enum HardhatConsoleCalls {

    log_0(log_0Call),
    log_1(log_1Call),
    log_2(log_2Call),
    log_3(log_3Call),
    log_4(log_4Call),
    log_5(log_5Call),
    log_6(log_6Call),
    log_7(log_7Call),
    log_8(log_8Call),
    log_9(log_9Call),
    log_10(log_10Call),
    log_11(log_11Call),
    log_12(log_12Call),
    log_13(log_13Call),
    log_14(log_14Call),
    log_15(log_15Call),
    log_16(log_16Call),
    log_17(log_17Call),
    log_18(log_18Call),
    log_19(log_19Call),
    log_20(log_20Call),
    log_21(log_21Call),
    log_22(log_22Call),
    log_23(log_23Call),
    log_24(log_24Call),
    log_25(log_25Call),
    log_26(log_26Call),
    log_27(log_27Call),
    log_28(log_28Call),
    log_29(log_29Call),
    log_30(log_30Call),
    log_31(log_31Call),
    log_32(log_32Call),
    log_33(log_33Call),
    log_34(log_34Call),
    log_35(log_35Call),
    log_36(log_36Call),
    log_37(log_37Call),
    log_38(log_38Call),
    log_39(log_39Call),
    log_40(log_40Call),
    log_41(log_41Call),
    log_42(log_42Call),
    log_43(log_43Call),
    log_44(log_44Call),
    log_45(log_45Call),
    log_46(log_46Call),
    log_47(log_47Call),
    log_48(log_48Call),
    log_49(log_49Call),
    log_50(log_50Call),
    log_51(log_51Call),
    log_52(log_52Call),
    log_53(log_53Call),
    log_54(log_54Call),
    log_55(log_55Call),
    log_56(log_56Call),
    log_57(log_57Call),
    log_58(log_58Call),
    log_59(log_59Call),
    log_60(log_60Call),
    log_61(log_61Call),
    log_62(log_62Call),
    log_63(log_63Call),
    log_64(log_64Call),
    log_65(log_65Call),
    log_66(log_66Call),
    log_67(log_67Call),
    log_68(log_68Call),
    log_69(log_69Call),
    log_70(log_70Call),
    log_71(log_71Call),
    log_72(log_72Call),
    log_73(log_73Call),
    log_74(log_74Call),
    log_75(log_75Call),
    log_76(log_76Call),
    log_77(log_77Call),
    log_78(log_78Call),
    log_79(log_79Call),
    log_80(log_80Call),
    log_81(log_81Call),
    log_82(log_82Call),
    log_83(log_83Call),
    log_84(log_84Call),
    log_85(log_85Call),
    log_86(log_86Call),
    log_87(log_87Call),
    log_88(log_88Call),
    log_89(log_89Call),
    log_90(log_90Call),
    log_91(log_91Call),
    log_92(log_92Call),
    log_93(log_93Call),
    log_94(log_94Call),
    log_95(log_95Call),
    log_96(log_96Call),
    log_97(log_97Call),
    log_98(log_98Call),
    log_99(log_99Call),
    log_100(log_100Call),
    log_101(log_101Call),
    log_102(log_102Call),
    log_103(log_103Call),
    log_104(log_104Call),
    log_105(log_105Call),
    log_106(log_106Call),
    log_107(log_107Call),
    log_108(log_108Call),
    log_109(log_109Call),
    log_110(log_110Call),
    log_111(log_111Call),
    log_112(log_112Call),
    log_113(log_113Call),
    log_114(log_114Call),
    log_115(log_115Call),
    log_116(log_116Call),
    log_117(log_117Call),
    log_118(log_118Call),
    log_119(log_119Call),
    log_120(log_120Call),
    log_121(log_121Call),
    log_122(log_122Call),
    log_123(log_123Call),
    log_124(log_124Call),
    log_125(log_125Call),
    log_126(log_126Call),
    log_127(log_127Call),
    log_128(log_128Call),
    log_129(log_129Call),
    log_130(log_130Call),
    log_131(log_131Call),
    log_132(log_132Call),
    log_133(log_133Call),
    log_134(log_134Call),
    log_135(log_135Call),
    log_136(log_136Call),
    log_137(log_137Call),
    log_138(log_138Call),
    log_139(log_139Call),
    log_140(log_140Call),
    log_141(log_141Call),
    log_142(log_142Call),
    log_143(log_143Call),
    log_144(log_144Call),
    log_145(log_145Call),
    log_146(log_146Call),
    log_147(log_147Call),
    log_148(log_148Call),
    log_149(log_149Call),
    log_150(log_150Call),
    log_151(log_151Call),
    log_152(log_152Call),
    log_153(log_153Call),
    log_154(log_154Call),
    log_155(log_155Call),
    log_156(log_156Call),
    log_157(log_157Call),
    log_158(log_158Call),
    log_159(log_159Call),
    log_160(log_160Call),
    log_161(log_161Call),
    log_162(log_162Call),
    log_163(log_163Call),
    log_164(log_164Call),
    log_165(log_165Call),
    log_166(log_166Call),
    log_167(log_167Call),
    log_168(log_168Call),
    log_169(log_169Call),
    log_170(log_170Call),
    log_171(log_171Call),
    log_172(log_172Call),
    log_173(log_173Call),
    log_174(log_174Call),
    log_175(log_175Call),
    log_176(log_176Call),
    log_177(log_177Call),
    log_178(log_178Call),
    log_179(log_179Call),
    log_180(log_180Call),
    log_181(log_181Call),
    log_182(log_182Call),
    log_183(log_183Call),
    log_184(log_184Call),
    log_185(log_185Call),
    log_186(log_186Call),
    log_187(log_187Call),
    log_188(log_188Call),
    log_189(log_189Call),
    log_190(log_190Call),
    log_191(log_191Call),
    log_192(log_192Call),
    log_193(log_193Call),
    log_194(log_194Call),
    log_195(log_195Call),
    log_196(log_196Call),
    log_197(log_197Call),
    log_198(log_198Call),
    log_199(log_199Call),
    log_200(log_200Call),
    log_201(log_201Call),
    log_202(log_202Call),
    log_203(log_203Call),
    log_204(log_204Call),
    log_205(log_205Call),
    log_206(log_206Call),
    log_207(log_207Call),
    log_208(log_208Call),
    log_209(log_209Call),
    log_210(log_210Call),
    log_211(log_211Call),
    log_212(log_212Call),
    log_213(log_213Call),
    log_214(log_214Call),
    log_215(log_215Call),
    log_216(log_216Call),
    log_217(log_217Call),
    log_218(log_218Call),
    log_219(log_219Call),
    log_220(log_220Call),
    log_221(log_221Call),
    log_222(log_222Call),
    log_223(log_223Call),
    log_224(log_224Call),
    log_225(log_225Call),
    log_226(log_226Call),
    log_227(log_227Call),
    log_228(log_228Call),
    log_229(log_229Call),
    log_230(log_230Call),
    log_231(log_231Call),
    log_232(log_232Call),
    log_233(log_233Call),
    log_234(log_234Call),
    log_235(log_235Call),
    log_236(log_236Call),
    log_237(log_237Call),
    log_238(log_238Call),
    log_239(log_239Call),
    log_240(log_240Call),
    log_241(log_241Call),
    log_242(log_242Call),
    log_243(log_243Call),
    log_244(log_244Call),
    log_245(log_245Call),
    log_246(log_246Call),
    log_247(log_247Call),
    log_248(log_248Call),
    log_249(log_249Call),
    log_250(log_250Call),
    log_251(log_251Call),
    log_252(log_252Call),
    log_253(log_253Call),
    log_254(log_254Call),
    log_255(log_255Call),
    log_256(log_256Call),
    log_257(log_257Call),
    log_258(log_258Call),
    log_259(log_259Call),
    log_260(log_260Call),
    log_261(log_261Call),
    log_262(log_262Call),
    log_263(log_263Call),
    log_264(log_264Call),
    log_265(log_265Call),
    log_266(log_266Call),
    log_267(log_267Call),
    log_268(log_268Call),
    log_269(log_269Call),
    log_270(log_270Call),
    log_271(log_271Call),
    log_272(log_272Call),
    log_273(log_273Call),
    log_274(log_274Call),
    log_275(log_275Call),
    log_276(log_276Call),
    log_277(log_277Call),
    log_278(log_278Call),
    log_279(log_279Call),
    log_280(log_280Call),
    log_281(log_281Call),
    log_282(log_282Call),
    log_283(log_283Call),
    log_284(log_284Call),
    log_285(log_285Call),
    log_286(log_286Call),
    log_287(log_287Call),
    log_288(log_288Call),
    log_289(log_289Call),
    log_290(log_290Call),
    log_291(log_291Call),
    log_292(log_292Call),
    log_293(log_293Call),
    log_294(log_294Call),
    log_295(log_295Call),
    log_296(log_296Call),
    log_297(log_297Call),
    log_298(log_298Call),
    log_299(log_299Call),
    log_300(log_300Call),
    log_301(log_301Call),
    log_302(log_302Call),
    log_303(log_303Call),
    log_304(log_304Call),
    log_305(log_305Call),
    log_306(log_306Call),
    log_307(log_307Call),
    log_308(log_308Call),
    log_309(log_309Call),
    log_310(log_310Call),
    log_311(log_311Call),
    log_312(log_312Call),
    log_313(log_313Call),
    log_314(log_314Call),
    log_315(log_315Call),
    log_316(log_316Call),
    log_317(log_317Call),
    log_318(log_318Call),
    log_319(log_319Call),
    log_320(log_320Call),
    log_321(log_321Call),
    log_322(log_322Call),
    log_323(log_323Call),
    log_324(log_324Call),
    log_325(log_325Call),
    log_326(log_326Call),
    log_327(log_327Call),
    log_328(log_328Call),
    log_329(log_329Call),
    log_330(log_330Call),
    log_331(log_331Call),
    log_332(log_332Call),
    log_333(log_333Call),
    log_334(log_334Call),
    log_335(log_335Call),
    log_336(log_336Call),
    log_337(log_337Call),
    log_338(log_338Call),
    log_339(log_339Call),
    log_340(log_340Call),
    log_341(log_341Call),
    log_342(log_342Call),
    logAddress(logAddressCall),
    logBool(logBoolCall),
    logBytes(logBytesCall),
    logBytes1(logBytes1Call),
    logBytes10(logBytes10Call),
    logBytes11(logBytes11Call),
    logBytes12(logBytes12Call),
    logBytes13(logBytes13Call),
    logBytes14(logBytes14Call),
    logBytes15(logBytes15Call),
    logBytes16(logBytes16Call),
    logBytes17(logBytes17Call),
    logBytes18(logBytes18Call),
    logBytes19(logBytes19Call),
    logBytes2(logBytes2Call),
    logBytes20(logBytes20Call),
    logBytes21(logBytes21Call),
    logBytes22(logBytes22Call),
    logBytes23(logBytes23Call),
    logBytes24(logBytes24Call),
    logBytes25(logBytes25Call),
    logBytes26(logBytes26Call),
    logBytes27(logBytes27Call),
    logBytes28(logBytes28Call),
    logBytes29(logBytes29Call),
    logBytes3(logBytes3Call),
    logBytes30(logBytes30Call),
    logBytes31(logBytes31Call),
    logBytes32(logBytes32Call),
    logBytes4(logBytes4Call),
    logBytes5(logBytes5Call),
    logBytes6(logBytes6Call),
    logBytes7(logBytes7Call),
    logBytes8(logBytes8Call),
    logBytes9(logBytes9Call),
    logInt(logIntCall),
    logString(logStringCall),
    logUint(logUintCall),

}

Container for all the HardhatConsole function calls.

Variants

log_0(log_0Call)

log_1(log_1Call)

log_2(log_2Call)

log_3(log_3Call)

log_4(log_4Call)

log_5(log_5Call)

log_6(log_6Call)

log_7(log_7Call)

log_8(log_8Call)

log_9(log_9Call)

log_10(log_10Call)

log_11(log_11Call)

log_12(log_12Call)

log_13(log_13Call)

log_14(log_14Call)

log_15(log_15Call)

log_16(log_16Call)

log_17(log_17Call)

log_18(log_18Call)

log_19(log_19Call)

log_20(log_20Call)

log_21(log_21Call)

log_22(log_22Call)

log_23(log_23Call)

log_24(log_24Call)

log_25(log_25Call)

log_26(log_26Call)

log_27(log_27Call)

log_28(log_28Call)

log_29(log_29Call)

log_30(log_30Call)

log_31(log_31Call)

log_32(log_32Call)

log_33(log_33Call)

log_34(log_34Call)

log_35(log_35Call)

log_36(log_36Call)

log_37(log_37Call)

log_38(log_38Call)

log_39(log_39Call)

log_40(log_40Call)

log_41(log_41Call)

log_42(log_42Call)

log_43(log_43Call)

log_44(log_44Call)

log_45(log_45Call)

log_46(log_46Call)

log_47(log_47Call)

log_48(log_48Call)

log_49(log_49Call)

log_50(log_50Call)

log_51(log_51Call)

log_52(log_52Call)

log_53(log_53Call)

log_54(log_54Call)

log_55(log_55Call)

log_56(log_56Call)

log_57(log_57Call)

log_58(log_58Call)

log_59(log_59Call)

log_60(log_60Call)

log_61(log_61Call)

log_62(log_62Call)

log_63(log_63Call)

log_64(log_64Call)

log_65(log_65Call)

log_66(log_66Call)

log_67(log_67Call)

log_68(log_68Call)

log_69(log_69Call)

log_70(log_70Call)

log_71(log_71Call)

log_72(log_72Call)

log_73(log_73Call)

log_74(log_74Call)

log_75(log_75Call)

log_76(log_76Call)

log_77(log_77Call)

log_78(log_78Call)

log_79(log_79Call)

log_80(log_80Call)

log_81(log_81Call)

log_82(log_82Call)

log_83(log_83Call)

log_84(log_84Call)

log_85(log_85Call)

log_86(log_86Call)

log_87(log_87Call)

log_88(log_88Call)

log_89(log_89Call)

log_90(log_90Call)

log_91(log_91Call)

log_92(log_92Call)

log_93(log_93Call)

log_94(log_94Call)

log_95(log_95Call)

log_96(log_96Call)

log_97(log_97Call)

log_98(log_98Call)

log_99(log_99Call)

log_100(log_100Call)

log_101(log_101Call)

log_102(log_102Call)

log_103(log_103Call)

log_104(log_104Call)

log_105(log_105Call)

log_106(log_106Call)

log_107(log_107Call)

log_108(log_108Call)

log_109(log_109Call)

log_110(log_110Call)

log_111(log_111Call)

log_112(log_112Call)

log_113(log_113Call)

log_114(log_114Call)

log_115(log_115Call)

log_116(log_116Call)

log_117(log_117Call)

log_118(log_118Call)

log_119(log_119Call)

log_120(log_120Call)

log_121(log_121Call)

log_122(log_122Call)

log_123(log_123Call)

log_124(log_124Call)

log_125(log_125Call)

log_126(log_126Call)

log_127(log_127Call)

log_128(log_128Call)

log_129(log_129Call)

log_130(log_130Call)

log_131(log_131Call)

log_132(log_132Call)

log_133(log_133Call)

log_134(log_134Call)

log_135(log_135Call)

log_136(log_136Call)

log_137(log_137Call)

log_138(log_138Call)

log_139(log_139Call)

log_140(log_140Call)

log_141(log_141Call)

log_142(log_142Call)

log_143(log_143Call)

log_144(log_144Call)

log_145(log_145Call)

log_146(log_146Call)

log_147(log_147Call)

log_148(log_148Call)

log_149(log_149Call)

log_150(log_150Call)

log_151(log_151Call)

log_152(log_152Call)

log_153(log_153Call)

log_154(log_154Call)

log_155(log_155Call)

log_156(log_156Call)

log_157(log_157Call)

log_158(log_158Call)

log_159(log_159Call)

log_160(log_160Call)

log_161(log_161Call)

log_162(log_162Call)

log_163(log_163Call)

log_164(log_164Call)

log_165(log_165Call)

log_166(log_166Call)

log_167(log_167Call)

log_168(log_168Call)

log_169(log_169Call)

log_170(log_170Call)

log_171(log_171Call)

log_172(log_172Call)

log_173(log_173Call)

log_174(log_174Call)

log_175(log_175Call)

log_176(log_176Call)

log_177(log_177Call)

log_178(log_178Call)

log_179(log_179Call)

log_180(log_180Call)

log_181(log_181Call)

log_182(log_182Call)

log_183(log_183Call)

log_184(log_184Call)

log_185(log_185Call)

log_186(log_186Call)

log_187(log_187Call)

log_188(log_188Call)

log_189(log_189Call)

log_190(log_190Call)

log_191(log_191Call)

log_192(log_192Call)

log_193(log_193Call)

log_194(log_194Call)

log_195(log_195Call)

log_196(log_196Call)

log_197(log_197Call)

log_198(log_198Call)

log_199(log_199Call)

log_200(log_200Call)

log_201(log_201Call)

log_202(log_202Call)

log_203(log_203Call)

log_204(log_204Call)

log_205(log_205Call)

log_206(log_206Call)

log_207(log_207Call)

log_208(log_208Call)

log_209(log_209Call)

log_210(log_210Call)

log_211(log_211Call)

log_212(log_212Call)

log_213(log_213Call)

log_214(log_214Call)

log_215(log_215Call)

log_216(log_216Call)

log_217(log_217Call)

log_218(log_218Call)

log_219(log_219Call)

log_220(log_220Call)

log_221(log_221Call)

log_222(log_222Call)

log_223(log_223Call)

log_224(log_224Call)

log_225(log_225Call)

log_226(log_226Call)

log_227(log_227Call)

log_228(log_228Call)

log_229(log_229Call)

log_230(log_230Call)

log_231(log_231Call)

log_232(log_232Call)

log_233(log_233Call)

log_234(log_234Call)

log_235(log_235Call)

log_236(log_236Call)

log_237(log_237Call)

log_238(log_238Call)

log_239(log_239Call)

log_240(log_240Call)

log_241(log_241Call)

log_242(log_242Call)

log_243(log_243Call)

log_244(log_244Call)

log_245(log_245Call)

log_246(log_246Call)

log_247(log_247Call)

log_248(log_248Call)

log_249(log_249Call)

log_250(log_250Call)

log_251(log_251Call)

log_252(log_252Call)

log_253(log_253Call)

log_254(log_254Call)

log_255(log_255Call)

log_256(log_256Call)

log_257(log_257Call)

log_258(log_258Call)

log_259(log_259Call)

log_260(log_260Call)

log_261(log_261Call)

log_262(log_262Call)

log_263(log_263Call)

log_264(log_264Call)

log_265(log_265Call)

log_266(log_266Call)

log_267(log_267Call)

log_268(log_268Call)

log_269(log_269Call)

log_270(log_270Call)

log_271(log_271Call)

log_272(log_272Call)

log_273(log_273Call)

log_274(log_274Call)

log_275(log_275Call)

log_276(log_276Call)

log_277(log_277Call)

log_278(log_278Call)

log_279(log_279Call)

log_280(log_280Call)

log_281(log_281Call)

log_282(log_282Call)

log_283(log_283Call)

log_284(log_284Call)

log_285(log_285Call)

log_286(log_286Call)

log_287(log_287Call)

log_288(log_288Call)

log_289(log_289Call)

log_290(log_290Call)

log_291(log_291Call)

log_292(log_292Call)

log_293(log_293Call)

log_294(log_294Call)

log_295(log_295Call)

log_296(log_296Call)

log_297(log_297Call)

log_298(log_298Call)

log_299(log_299Call)

log_300(log_300Call)

log_301(log_301Call)

log_302(log_302Call)

log_303(log_303Call)

log_304(log_304Call)

log_305(log_305Call)

log_306(log_306Call)

log_307(log_307Call)

log_308(log_308Call)

log_309(log_309Call)

log_310(log_310Call)

log_311(log_311Call)

log_312(log_312Call)

log_313(log_313Call)

log_314(log_314Call)

log_315(log_315Call)

log_316(log_316Call)

log_317(log_317Call)

log_318(log_318Call)

log_319(log_319Call)

log_320(log_320Call)

log_321(log_321Call)

log_322(log_322Call)

log_323(log_323Call)

log_324(log_324Call)

log_325(log_325Call)

log_326(log_326Call)

log_327(log_327Call)

log_328(log_328Call)

log_329(log_329Call)

log_330(log_330Call)

log_331(log_331Call)

log_332(log_332Call)

log_333(log_333Call)

log_334(log_334Call)

log_335(log_335Call)

log_336(log_336Call)

log_337(log_337Call)

log_338(log_338Call)

log_339(log_339Call)

log_340(log_340Call)

log_341(log_341Call)

log_342(log_342Call)

logAddress(logAddressCall)

logBool(logBoolCall)

logBytes(logBytesCall)

logBytes1(logBytes1Call)

logBytes10(logBytes10Call)

logBytes11(logBytes11Call)

logBytes12(logBytes12Call)

logBytes13(logBytes13Call)

logBytes14(logBytes14Call)

logBytes15(logBytes15Call)

logBytes16(logBytes16Call)

logBytes17(logBytes17Call)

logBytes18(logBytes18Call)

logBytes19(logBytes19Call)

logBytes2(logBytes2Call)

logBytes20(logBytes20Call)

logBytes21(logBytes21Call)

logBytes22(logBytes22Call)

logBytes23(logBytes23Call)

logBytes24(logBytes24Call)

logBytes25(logBytes25Call)

logBytes26(logBytes26Call)

logBytes27(logBytes27Call)

logBytes28(logBytes28Call)

logBytes29(logBytes29Call)

logBytes3(logBytes3Call)

logBytes30(logBytes30Call)

logBytes31(logBytes31Call)

logBytes32(logBytes32Call)

logBytes4(logBytes4Call)

logBytes5(logBytes5Call)

logBytes6(logBytes6Call)

logBytes7(logBytes7Call)

logBytes8(logBytes8Call)

logBytes9(logBytes9Call)

logInt(logIntCall)

logString(logStringCall)

logUint(logUintCall)

Implementations

impl HardhatConsoleCalls

pub const SELECTORS: &'static [[u8; 4]] = _

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.