Module foundry_evm_core::abi::hardhat_console

Modules

HardhatConsole

Generated by the following Solidity interface…

Statics

HARDHAT_CONSOLE_SELECTOR_PATCHES

Maps all the hardhat/console.log log selectors that use the legacy ABI (int, uint) to their normalized counterparts (int256, uint256).

Functions

hh_console_selector

Returns the ABI-normalized selector for the given Hardhat console function selector.

patch_hh_console_selector

Patches the given Hardhat console function selector to its ABI-normalized form.