foundry_evm_core::backend

Struct Backend

pub struct Backend {
    forks: MultiFork,
    mem_db: FoundryEvmInMemoryDB,
    fork_init_journaled_state: JournaledState,
    active_fork_ids: Option<(LocalForkId, usize)>,
    inner: BackendInner,
}

Provides the underlying revm::Database implementation.

A Backend can be initialised in two forms:

1. Empty in-memory Database

This is the default variant: an empty revm::Database

2. Forked Database

A revm::Database that forks off a remote client

In addition to that we support forking manually on the fly. Additional forks can be created. Each unique fork is identified by its unique ForkId. We treat forks as unique if they have the same (endpoint, block number) pair.

When it comes to testing, it’s intended that each contract will use its own Backend (Backend::clone). This way each contract uses its own encapsulated evm state. For in-memory testing, the database is just an owned revm::InMemoryDB.

Each Fork, identified by a unique id, uses completely separate storage, write operations are performed only in the fork’s own database, ForkDB.

A ForkDB consists of 2 halves:

everything fetched from the remote is readonly

all local changes (instructed by the contract) are written to the backend’s db and don’t alter the state of the remote client.

Fork swapping

Multiple “forks” can be created Backend::create_fork(), however only 1 can be used by the db. However, their state can be hot-swapped by swapping the read half of db from one fork to another. When swapping forks (Backend::select_fork()) we also update the current Env of the EVM accordingly, so that all block.* config values match

When another for is selected DatabaseExt::select_fork() the entire storage, including JournaledState is swapped, but the storage of the caller’s and the test contract account is always cloned. This way a fork has entirely separate storage but data can still be shared across fork boundaries via stack and contract variables.

Snapshotting

A snapshot of the current overall state can be taken at any point in time. A snapshot is identified by a unique id that’s returned when a snapshot is created. A snapshot can only be reverted once. After a successful revert, the same snapshot id cannot be used again. Reverting a snapshot replaces the current active state with the snapshot state, the snapshot is deleted afterwards, as well as any snapshots taken after the reverted snapshot, (e.g.: reverting to id 0x1 will delete snapshots with ids 0x1, 0x2, etc.)

Note: State snapshots work across fork-swaps, e.g. if fork A is currently active, then a snapshot is created before fork B is selected, then fork A will be the active fork again after reverting the snapshot.

Fields
forks: MultiFork

The access point for managing forks

mem_db: FoundryEvmInMemoryDBfork_init_journaled_state: JournaledState

The journaled_state to use to initialize new forks with

The way JournaledState works is, that it holds the “hot” accounts loaded from the underlying Database that feeds the Account and State data to the journaled_state so it can apply changes to the state while the EVM executes.

In a way the JournaledState is something like a cache that

check if account is already loaded (hot)

if not load from the Database (this will then retrieve the account via RPC in forking mode)

To properly initialize we store the JournaledState before the first fork is selected (DatabaseExt::select_fork).

This will be an empty JournaledState, which will be populated with persistent accounts, See Self::update_fork_db().

active_fork_ids: Option<(LocalForkId, usize)>

The currently active fork database

If this is set, then the Backend is currently in forking mode

inner: BackendInner

holds additional Backend data

Implementations

impl Backend

pub fn spawn(fork: Option<CreateFork>) -> Result<Self>

Creates a new Backend with a spawned multi fork thread.

If fork is Some this will use a fork database, otherwise with an in-memory database.

pub fn new(forks: MultiFork, fork: Option<CreateFork>) -> Result<Self>

Creates a new instance of Backend

If fork is Some this will use a fork database, otherwise with an in-memory database.

Prefer using spawn instead.

pub(crate) fn new_with_fork( id: &ForkId, fork: Fork, journaled_state: JournaledState, ) -> Result<Self>

Creates a new instance of Backend with fork added to the fork database and sets the fork as active

pub fn clone_empty(&self) -> Self

Creates a new instance with a BackendDatabase::InMemory cache layer for the CacheDB

pub fn insert_account_info(&mut self, address: Address, account: AccountInfo)

pub fn insert_account_storage( &mut self, address: Address, slot: U256, value: U256, ) -> Result<(), DatabaseError>

Inserts a value on an account’s storage without overriding account info

pub fn replace_account_storage( &mut self, address: Address, storage: Map<U256, U256>, ) -> Result<(), DatabaseError>

Completely replace an account’s storage without overriding account info.

When forking, this causes the backend to assume a 0 value for all unset storage slots instead of trying to fetch it.

pub fn state_snapshots( &self, ) -> &StateSnapshots<BackendStateSnapshot<BackendDatabaseSnapshot>>

Returns all snapshots created in this backend

pub fn set_test_contract(&mut self, acc: Address) -> &mut Self

Sets the address of the DSTest contract that is being executed

This will also mark the caller as persistent and remove the persistent status from the previous test contract address

This will also grant cheatcode access to the test account

pub fn set_caller(&mut self, acc: Address) -> &mut Self

Sets the caller address

pub fn set_spec_id(&mut self, spec_id: SpecId) -> &mut Self

Sets the current spec id

pub fn caller_address(&self) -> Option<Address>

Returns the set caller address

pub fn has_state_snapshot_failure(&self) -> bool

Failures occurred in state snapshots are tracked when the state snapshot is reverted.

If an error occurs in a restored state snapshot, the test is considered failed.

This returns whether there was a reverted state snapshot that recorded an error.

pub fn set_state_snapshot_failure(&mut self, has_state_snapshot_failure: bool)

Sets the state snapshot failure flag.

pub(crate) fn update_fork_db( &self, active_journaled_state: &mut JournaledState, target_fork: &mut Fork, )

When creating or switching forks, we update the AccountInfo of the contract

pub(crate) fn update_fork_db_contracts( &self, accounts: impl IntoIterator<Item = Address>, active_journaled_state: &mut JournaledState, target_fork: &mut Fork, )

Merges the state of all accounts from the currently active db into the given fork

pub fn mem_db(&self) -> &FoundryEvmInMemoryDB

Returns the memory db used if not in forking mode

pub fn is_active_fork(&self, id: LocalForkId) -> bool

Returns true if the id is currently active

pub fn is_in_forking_mode(&self) -> bool

Returns true if the Backend is currently in forking mode

pub fn active_fork(&self) -> Option<&Fork>

Returns the currently active Fork, if any

pub fn active_fork_mut(&mut self) -> Option<&mut Fork>

Returns the currently active Fork, if any

pub fn active_fork_db(&self) -> Option<&CacheDB<SharedBackend>>

Returns the currently active ForkDB, if any

pub fn active_fork_db_mut(&mut self) -> Option<&mut CacheDB<SharedBackend>>

Returns the currently active ForkDB, if any

pub fn db(&self) -> &dyn Database<Error = DatabaseError>

Returns the current database implementation as a &dyn value.

pub fn db_mut(&mut self) -> &mut dyn Database<Error = DatabaseError>

Returns the current database implementation as a &mut dyn value.

pub(crate) fn create_db_snapshot(&self) -> BackendDatabaseSnapshot

Creates a snapshot of the currently active database

pub fn merged_logs(&self, logs: Vec<Log>) -> Vec<Log>

Since each Fork tracks logs separately, we need to merge them to get all of them

pub(crate) fn initialize(&mut self, env: &Env)

Initializes settings we need to keep track of.

We need to track these mainly to prevent issues when switching between different evms

pub fn inspect<I: InspectorExt>( &mut self, env: &mut Env, inspector: I, ) -> Result<ResultAndState>

Executes the configured test call of the env without committing state changes.

Note: in case there are any cheatcodes executed that modify the environment, this will update the given env with the new values.

pub fn is_existing_precompile(&self, addr: &Address) -> bool

Returns true if the address is a precompile

fn set_init_journaled_state(&mut self, journaled_state: JournaledState)

Sets the initial journaled state to use when initializing forks

fn prepare_init_journal_state(&mut self) -> Result<(), BackendError>

Cleans up already loaded accounts that would be initialized without the correct data from the fork.

It can happen that an account is loaded before the first fork is selected, like getNonce(addr), which will load an empty account by default.

This account data then would not match the account data of a fork if it exists. So when the first fork is initialized we replace these accounts with the actual account as it exists on the fork.

fn get_block_number_and_block_for_transaction( &self, id: LocalForkId, transaction: B256, ) -> Result<(u64, AnyRpcBlock)>

Returns the block numbers required for replaying a transaction

pub fn replay_until( &mut self, id: LocalForkId, env: Env, tx_hash: B256, journaled_state: &mut JournaledState, ) -> Result<Option<Transaction<AnyTxEnvelope>>>

Replays all the transactions at the forks current block that were mined before the tx

Returns the unmined transaction that corresponds to the given tx_hash