foundry_evm_core::backend

Trait BackendStrategyRunner

pub trait BackendStrategyRunner:
    Debug
    + Send
    + Sync {
    // Required methods
    fn inspect(
        &self,
        backend: &mut Backend,
        env: &mut EnvWithHandlerCfg,
        inspector: &mut dyn InspectorExt,
        inspect_ctx: Box<dyn Any>,
    ) -> Result<ResultAndState>;
    fn update_fork_db(
        &self,
        ctx: &mut dyn BackendStrategyContext,
        active_fork: Option<&Fork>,
        mem_db: &FoundryEvmInMemoryDB,
        backend_inner: &BackendInner,
        active_journaled_state: &mut JournaledState,
        target_fork: &mut Fork,
    );
    fn merge_journaled_state_data(
        &self,
        ctx: &mut dyn BackendStrategyContext,
        addr: Address,
        active_journaled_state: &JournaledState,
        fork_journaled_state: &mut JournaledState,
    );
    fn merge_db_account_data(
        &self,
        ctx: &mut dyn BackendStrategyContext,
        addr: Address,
        active: &ForkDB,
        fork_db: &mut ForkDB,
    );
}

Stateless strategy runner for BackendStrategy.

Required Methods

fn inspect( &self, backend: &mut Backend, env: &mut EnvWithHandlerCfg, inspector: &mut dyn InspectorExt, inspect_ctx: Box<dyn Any>, ) -> Result<ResultAndState>

Executes the configured test call of the env without committing state changes.

fn update_fork_db( &self, ctx: &mut dyn BackendStrategyContext, active_fork: Option<&Fork>, mem_db: &FoundryEvmInMemoryDB, backend_inner: &BackendInner, active_journaled_state: &mut JournaledState, target_fork: &mut Fork, )

When creating or switching forks, we update the AccountInfo of the contract

fn merge_journaled_state_data( &self, ctx: &mut dyn BackendStrategyContext, addr: Address, active_journaled_state: &JournaledState, fork_journaled_state: &mut JournaledState, )

Clones the account data from the active_journaled_state into the fork_journaled_state

fn merge_db_account_data( &self, ctx: &mut dyn BackendStrategyContext, addr: Address, active: &ForkDB, fork_db: &mut ForkDB, )

Clones the account data from the active db into the ForkDB