foundry_evm_core

Module decode

Various utilities to decode test results.

Structs

RevertDecoder

Decodes revert data.

SkipReason

A skip reason.

Functions

decode_as_non_empty_string 🔒

Helper function that decodes provided error as an ABI encoded or an ASCII string (if not empty).

decode_console_log

Decode a single log.

decode_console_logs

Decode a set of logs, only returning logs from DSTest logging events and Hardhat’s console.log

trimmed_hex 🔒