Module foundry_evm_core::fork::backend

Smart caching and deduplication of requests when using a forking provider

Structs

BackendHandler

Handles an internal provider and listens for requests.

SharedBackend

A cloneable backend type that shares access to the backend data with all its clones.

Enums

BackendRequest 🔒

The Request type the Backend listens for

ProviderRequest 🔒

Request variants that are executed by the provider

Type Aliases

AccountFuture 🔒

AccountInfoSender 🔒

BlockHashFuture 🔒

BlockHashSender 🔒

FullBlockFuture 🔒

FullBlockSender 🔒

StorageFuture 🔒

StorageSender 🔒

TransactionFuture 🔒

TransactionSender 🔒