Module foundry_evm_core::fork::multi

Support for running multiple fork backends

The design is similar to the single SharedBackend, BackendHandler but supports multiple concurrently active pairs at once.

Structs

CreatedFork 🔒

Tracks the created Fork

ForkId

The unique identifier for a specific fork, this could be the name of the network a custom descriptive name.

MultiFork

The Sender half of multi fork pair. Can send requests to the MultiForkHandler to create forks

MultiForkHandler

The type that manages connections in the background

ShutDownMultiFork 🔒

A type that’s used to signaling the MultiForkHandler when it’s time to shut down.

Enums

ForkTask 🔒

Request 🔒

Request that’s send to the handler

Functions

create_fork 🔒

Creates a new fork

Type Aliases

CreateFuture 🔒

CreateSender 🔒

GetEnvSender 🔒

Handler 🔒