foundry_evm_core

Trait InspectorExt

pub trait InspectorExt: for<'a> Inspector<EthEvmContext<&'a mut dyn DatabaseExt>> {
    // Provided methods
    fn should_use_create2_factory(
        &mut self,
        _context: &mut EthEvmContext<&mut dyn DatabaseExt>,
        _inputs: &CreateInputs,
    ) -> bool { ... }
    fn console_log(&mut self, msg: &str) { ... }
    fn get_networks(&self) -> NetworkConfigs { ... }
    fn create2_deployer(&self) -> Address { ... }
    fn trace_revive(
        &mut self,
        _context: Ecx<'_, '_, '_>,
        _call_traces: Box<dyn Any>,
        _record_top_call: bool,
    ) { ... }
}

An extension trait that allows us to add additional hooks to Inspector for later use in handlers.

Provided Methods

fn should_use_create2_factory( &mut self, _context: &mut EthEvmContext<&mut dyn DatabaseExt>, _inputs: &CreateInputs, ) -> bool

Determines whether the DEFAULT_CREATE2_DEPLOYER should be used for a CREATE2 frame.

If this function returns true, we’ll replace CREATE2 frame with a CALL frame to CREATE2 factory.

fn console_log(&mut self, msg: &str)

Simulates console.log invocation.

fn get_networks(&self) -> NetworkConfigs

Returns configured networks.

fn create2_deployer(&self) -> Address

Returns the CREATE2 deployer address.

fn trace_revive( &mut self, _context: Ecx<'_, '_, '_>, _call_traces: Box<dyn Any>, _record_top_call: bool, )

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".