Crate foundry_evm_core

foundry-evm-core

Core EVM abstractions.

Re-exports

pub use env::*;

Modules

abi

backend

Foundry’s main executor backend abstraction and implementation.

buffer

constants

decode

Various utilities to decode test results.

either_evm

env

evm

fork

ic

opts

precompiles

state_snapshot

Support for snapshotting different states

utils

Traits

InspectorExt

An extension trait that allows us to add additional hooks to Inspector for later use in handlers.