Crate foundry_evm_core

foundry-evm-core

Core EVM abstractions.

Modules

abi

backend

Foundry’s main executor backend abstraction and implementation.

constants

decode

Various utilities to decode test results.

fork

ic 🔒

opcodes

Opcode utils

opts

snapshot

support for snapshotting different states

utils

Traits

InspectorExt

An extension trait that allows us to add additional hooks to Inspector for later use in handlers.