foundry_evm_core

Module utils

Functions

apply_chain_and_block_specific_env_changes

Depending on the configured chain id and block number this should apply any specific changes

cold_path

Hints to the compiler that this is a cold path, i.e. unlikely to be taken.

configure_tx_env

Configures the env for the given RPC transaction. Accounts for an impersonated transaction by resetting the env.tx.caller field to tx.from.

configure_tx_req_env

Configures the env for the given RPC transaction request. impersonated_from is the address of the impersonated account. This helps account for an impersonated transaction by resetting the env.tx.caller field to impersonated_from.

get_blob_base_fee_update_fraction

Derive the blob base fee update fraction based on the chain and timestamp by checking the hardfork.

get_blob_base_fee_update_fraction_by_spec_id

Returns the blob base fee update fraction based on the spec id.

get_function

Given an ABI and selector, it tries to find the respective function.

Type Aliases

StateChangeset

EVM State is a mapping from addresses to accounts.